tool/dbgenerator/requirements.txt
mysql-connector-python==8.2.0
jinja2==3.1.2
